tcbasic implements a small subset of BASIC known as Tiny BASIC. It provides the following statements and commands: INPUT, PRINT, LET, GOTO, GOSUB, RETURN, IF, END, CLEAR, LIST, and RUN. Integer arithmetic is supported, and strings may be PRINTed. A built-in RND(n) function provides random numbers.
The small size of the language make it easy to learn and master
while providing all of the building blocks needed to develop many
interesting programs. tcbasic runs on a variety of platforms and
aims to be as portable as possible.

WWW: https://github.com/tcort/tcbasic

PR:		197938
Submitted by:	Thomas Cort <linuxgeek@gmail.com>