Unbreak graphics/inkscape build on FreeBSD 9.3 i386.
The base version of clang 3.4.1 in FreeBSD 9.3 is missing some
patches that are present in clang 3.4.1 in FreeBSD 10.1.  One of
these patches appears to fix a code generation bug on i386 that is
triggered when building graphics/inkscape.

Work around this issue by building inkscape with lang/clang34 from
ports on FreeBSD 9.3 i386.

Approved by:	mat (mentor)
MFH:		2015Q2
Differential Revision:	https://reviews.freebsd.org/D2417