Set LIBLZMA_INCLUDE_DIR and LIBLZMA_LIBRARY in the initial cache.
Do it so that we can hardcode these variables to point to base's liblzma,
otherwise the liblzma.so installed by archivers/lzmalib will be chosen if
present and the configuration process will fail.

This is done only so that `make configure' works, as liblzma is only used if
we build CMake's bundled libarchive.

MFH:		2015Q2