Allow the build of the AFC backend for gvfs.
The AFC backend for non mass-storage Apple devices like
the iPhones and such, those use a custom usb protocol like it happens with
mtp devices. Do not enable it by default.

Do not force the rebuild of the port because the new option is
disabled by default. Bumping portrevision would surely force the
rebuild of many dependant ports, at least poudriere would do that. Keeping
the portversion would allow to save the rebuild if the option is being kept
as disabled.