mail/milter-bogom: Fix build on systems without base sendmail
There was an attempt to support ports sendmail made, but the attempt
didn't work.  Well, it accidently worked because the compilers looked
for headers and libraries at LOCALBASE/include and LOCALBASE/lib so
the bad LDFLAGS value was ignored.  Fix ports-sendmail LDFLAGS and
explicitly set -I so milter-bogom builds on all compilers.

Approved by:	just fix it